am: move IDebugFunctions
