+++ src/core/hle/service/am/service/debug_functions.cpp
@@ -1,8 +1,7 @@
 // SPDX-FileCopyrightText: Copyright 2024 yuzu Emulator Project
 // SPDX-License-Identifier: GPL-2.0-or-later
 
-#include "core/hle/service/am/debug_functions.h"
-#include "core/hle/service/ipc_helpers.h"
+#include "core/hle/service/am/service/debug_functions.h"
 
 namespace Service::AM {
