+++ src/core/hle/service/am/service/debug_functions.h
